added render function and localStorage in a different module and cleaned up a bit
