+++ src/index.js
@@ -1,6 +1,6 @@
 import { DomManipulation } from "/src/domManipulation.js"
-import { addBookToLibrary, render, createBook } from "/src/addingBooks.js"
-//import { render } from "/src/renderBook.js"
+import { addBookToLibrary } from "/src/addingBooks.js"
+import { createBook } from "/src/renderBook.js"
 import {myLibrary} from "/src/addingBooks.js"
 
 
@@ -11,7 +11,6 @@ const addBookButton = document.querySelector('.add-book')
 addBookButton.addEventListener('click', () => {
 
     addBookToLibrary()
-    // render()
     createBook();
     console.log(myLibrary)
 });
